audit.go: move types to separate file
No functional change.
